+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add changes here.
 
+## [0.0.2] - 2025-11-01
+
+### Added
+
+- Topbar made with shadcn/ui
+- Placeholders for body and footer
+
 ## [0.0.1] - 2025-10-31
 
 ### Added